tests/test_extractors.py
import pytest
from shopsmart.tools.scrapers import extract_product

def test_extract_handles_unknown():
    # Should not raise; None or Product
    url = "https://example.com/product"
    try:
        p = extract_product(url)  # likely None; just ensure no crash
    except Exception as e:
        pytest.fail(f"Extractor crashed: {e}")
