requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
extruct>=0.16.0
w3lib>=2.2.1
rapidfuzz>=3.9.3
tavily-python>=0.3.3
pydantic>=2.7.1
pandas>=2.2.2
matplotlib>=3.8.4
crewai>=0.100.1
crewai-tools>=0.33.0
litellm>=1.50.0
tenacity>=8.3.0
urllib3>=2.2.2

streamlit-webrtc
SpeechRecognition
